+++ ficary/__init__.py
@@ -2,7 +2,7 @@
 
 import logging as _logging
 
-__version__ = "2.14.1"
+__version__ = "2.14.2"
 
 _logger = _logging.getLogger(__name__)